apps/auth-service/requirements.txt
fastapi
uvicorn[standard]
redis
psycopg2-binary
pydantic
python-multipart
sqlalchemy
python-jose[cryptography]
passlib[argon2]
argon2-cffi
